+++ src/entry/types.ts
@@ -1,5 +1,7 @@
 import { Descendant } from "slate"
 
+export type OnImageChangeHandler = (file: File) => Promise<string>
+
 export type WysimarkEditor = {
   /**
    * Private state for the wysimark editor.
@@ -21,6 +23,11 @@ export type WysimarkEditor = {
      * Function to toggle Raw mode
      */
     toggleRawMode?: () => void
+
+    /**
+     * Handler for image file upload
+     */
+    onImageChange?: OnImageChangeHandler
   }
   /**
    * Public methods for the wysimark editor.