+++ src/utils/translations.ts
@@ -38,6 +38,11 @@ interface Translations {
     switchToRawMarkdown: string;
     codeBlock: string;
     increaseQuoteDepth: string;
+    register: string;
+    imageSourceUrl: string;
+    imageSourceFile: string;
+    selectFile: string;
+    uploading: string;
   };
 }
 
@@ -81,6 +86,11 @@ export const translations: Translations = {
     switchToRawMarkdown: "マークダウン表示に切り替え",
     codeBlock: "コードブロック",
     increaseQuoteDepth: "引用を重ねる",
+    register: "登録",
+    imageSourceUrl: "URL",
+    imageSourceFile: "ファイル",
+    selectFile: "ファイルを選択",
+    uploading: "アップロード中...",
   },
   en: {
     bold: "Bold",
@@ -121,6 +131,11 @@ export const translations: Translations = {
     switchToRawMarkdown: "Switch to raw markdown",
     codeBlock: "Code Block",
     increaseQuoteDepth: "Increase Quote Depth",
+    register: "Register",
+    imageSourceUrl: "URL",
+    imageSourceFile: "File",
+    selectFile: "Select File",
+    uploading: "Uploading...",
   },
 };